+++ provenance_lib/click_commands.py
@@ -3,7 +3,7 @@ import os
 
 from .parse import ProvDAG
 from .replay import replay_fp, write_citations
-from .usage_drivers import DRIVER_CHOICES, DRIVER_NAMES
+from ._usage_drivers import DRIVER_CHOICES, DRIVER_NAMES
 from .util import FileName
