+++ provenance_lib/replay.py
@@ -11,7 +11,7 @@ from typing import Dict, Iterator, List, Optional, Set
 
 from ._archive_parser import ProvNode
 from .parse import ProvDAG, UUID
-from .usage_drivers import DRIVER_CHOICES, SUPPORTED_USAGE_DRIVERS, Usage
+from ._usage_drivers import DRIVER_CHOICES, SUPPORTED_USAGE_DRIVERS, Usage
 from .util import FileName, camel_to_snake
 from .yaml_constructors import MetadataInfo
